requirements-ci.txt
-r requirements.txt

black==22.6.0
django-stubs==1.12.0
djangorestframework-stubs==1.7.0
flake8==4.0.1
isort==5.10.1
mypy==0.961
pre-commit==2.20.0
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-xdist==2.5.0
types-python-dateutil==2.8.18
